# contracts/management/v1/management-v1.openapi.json
{
  "components": {
    "schemas": {
      "CompanionCreateRequest": {
        "additionalProperties": false,
        "description": "Ask for another Eidolon.\n\n``operation_id`` is the client's, and it is what makes asking twice safe:\nevery identifier the Host derives comes from it, so a retry addresses the\nsame Eidolon rather than creating a second one. A client that generates a\nfresh id per attempt has opted out of that protection, which is why it is\nrequired rather than defaulted here.",
        "properties": {
          "display_name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "kind": {
            "default": "conversational",
            "maxLength": 32,
            "minLength": 1,
            "title": "Kind",
            "type": "string"
          },
          "operation_id": {
            "maxLength": 36,
            "minLength": 36,
            "title": "Operation Id",
            "type": "string"
          }
        },
        "required": [
          "operation_id",
          "display_name"
        ],
        "title": "CompanionCreateRequest",
        "type": "object"
      },
      "CompanionCreatedView": {
        "additionalProperties": false,
        "description": "The Eidolon that now exists, and whether its memory is up yet.",
        "properties": {
          "companion_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Companion Id",
            "type": "string"
          },
          "contract_version": {
            "const": "1",
            "default": "1",
            "title": "Contract Version",
            "type": "string"
          },
          "created": {
            "title": "Created",
            "type": "boolean"
          },
          "display_name": {
            "default": "",
            "maxLength": 128,
            "title": "Display Name",
            "type": "string"
          },
          "kind": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Kind",
            "type": "string"
          },
          "lifecycle_state": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Lifecycle State",
            "type": "string"
          },
          "memory_ready": {
            "title": "Memory Ready",
            "type": "boolean"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "companion_id",
          "kind",
          "lifecycle_state",
          "revision",
          "created",
          "memory_ready"
        ],
        "title": "CompanionCreatedView",
        "type": "object"
      },
      "CompanionDetailView": {
        "additionalProperties": false,
        "description": "One Eidolon, opened.\n\n``is_default`` is here and *not* on a roster row on purpose. A single answer\nabout a single Companion is a comparison the Host just made against the\nOwner's one pointer; a flag repeated across a list would let two rows claim\nit. Same fact, and only one shape of it can be self-contradictory.",
        "properties": {
          "companion_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Companion Id",
            "type": "string"
          },
          "contract_version": {
            "const": "1",
            "default": "1",
            "title": "Contract Version",
            "type": "string"
          },
          "display_name": {
            "default": "",
            "maxLength": 128,
            "title": "Display Name",
            "type": "string"
          },
          "is_default": {
            "title": "Is Default",
            "type": "boolean"
          },
          "kind": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Kind",
            "type": "string"
          },
          "lifecycle_state": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Lifecycle State",
            "type": "string"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "companion_id",
          "kind",
          "lifecycle_state",
          "revision",
          "is_default"
        ],
        "title": "CompanionDetailView",
        "type": "object"
      },
      "CompanionRosterView": {
        "additionalProperties": false,
        "description": "One page of this Owner's Eidolons, oldest first.",
        "properties": {
          "companions": {
            "items": {
              "$ref": "#/components/schemas/CompanionSummaryView"
            },
            "title": "Companions",
            "type": "array"
          },
          "contract_version": {
            "const": "1",
            "default": "1",
            "title": "Contract Version",
            "type": "string"
          },
          "default_companion_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Companion Id"
          },
          "next_cursor": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "companions"
        ],
        "title": "CompanionRosterView",
        "type": "object"
      },
      "CompanionSummaryView": {
        "additionalProperties": false,
        "description": "One Eidolon, as a list shows it.\n\nNo \"is the default\" flag. The page says which one is default exactly once,\nso a client cannot render two rows both claiming it, and cannot disagree\nwith the Owner record about which one it is.",
        "properties": {
          "companion_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Companion Id",
            "type": "string"
          },
          "created_at": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "default": "",
            "maxLength": 128,
            "title": "Display Name",
            "type": "string"
          },
          "kind": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Kind",
            "type": "string"
          },
          "lifecycle_state": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Lifecycle State",
            "type": "string"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "updated_at": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "companion_id",
          "kind",
          "lifecycle_state",
          "revision",
          "created_at",
          "updated_at"
        ],
        "title": "CompanionSummaryView",
        "type": "object"
      },
      "DefaultCompanionRequest": {
        "additionalProperties": false,
        "description": "Which of my Eidolons answers when I did not say which.\n\n``expected_revision`` is not ceremony a client can skip: it is the version\nof the Owner it last read, and it is what makes two phones (or a phone and a\nbrowser) unable to both win. A client that has not read the Owner has no\nbusiness changing this.",
        "properties": {
          "companion_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Companion Id",
            "type": "string"
          },
          "expected_revision": {
            "minimum": 1.0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "companion_id",
          "expected_revision"
        ],
        "title": "DefaultCompanionRequest",
        "type": "object"
      },
      "DefaultCompanionView": {
        "additionalProperties": false,
        "description": "Where the pointer now points, read back from the authority.",
        "properties": {
          "contract_version": {
            "const": "1",
            "default": "1",
            "title": "Contract Version",
            "type": "string"
          },
          "default_companion_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Companion Id"
          }
        },
        "title": "DefaultCompanionView",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ManagementContextView": {
        "additionalProperties": false,
        "description": "What a client reads before it draws anything.\n\n``capabilities`` is discovery, not permission: true means this Host can do\nthe thing at all, and whether this Controller may is answered per action. A\nname absent from the map is one this client has never heard of — a version\nskew — while a name present and false is a feature this Host cannot do yet.\n\n``limits`` values may be null, and a client must not substitute a number of\nits own: a limit nobody has measured is not a limit.",
        "properties": {
          "capabilities": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Capabilities",
            "type": "object"
          },
          "contract_version": {
            "const": "1",
            "default": "1",
            "title": "Contract Version",
            "type": "string"
          },
          "default_companion_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Companion Id"
          },
          "limits": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Limits",
            "type": "object"
          },
          "owner": {
            "$ref": "#/components/schemas/OwnerContextView"
          }
        },
        "required": [
          "owner",
          "capabilities",
          "limits"
        ],
        "title": "ManagementContextView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The single Owner management ABI. Both management clients are generated from this document; no route here accepts an owner_id, because the Owner comes from the authenticated Controller session.",
    "title": "Eidolon Owner Management",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/management/v1/companions": {
      "get": {
        "description": "This Owner's Eidolons.\n\nThere is no ``owner_id`` parameter and there will not be one: the Owner\nis whoever this session authenticated. ``cursor`` is the one thing a\nclient may vary, and it is a value this boundary handed it.",
        "operationId": "list_companions_api_management_v1_companions_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionRosterView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Companions",
        "tags": [
          "management"
        ]
      },
      "put": {
        "description": "Add another Eidolon for whoever is signed in.\n\nPUT with the client's operation id in the body rather than POST: the\nrequest states an end (\"this operation has produced an Eidolon\"), so a\nphone that lost the answer asks again and gets the same one. A POST here\nwould make a second Eidolon the normal consequence of a dropped\nresponse.",
        "operationId": "create_companion_api_management_v1_companions_put",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionCreatedView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Companion",
        "tags": [
          "management"
        ]
      }
    },
    "/api/management/v1/companions/{companion_id}": {
      "get": {
        "description": "One of this Owner's Eidolons.\n\nThe Companion is in the path and the Owner is not expressible, so the\npair is \"this id, for whoever is signed in\". One belonging to someone\nelse answers 404, not 403: an id that answers differently can be probed\nfor existence.",
        "operationId": "get_companion_api_management_v1_companions__companion_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "companion_id",
            "required": true,
            "schema": {
              "title": "Companion Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionDetailView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Companion",
        "tags": [
          "management"
        ]
      }
    },
    "/api/management/v1/context": {
      "get": {
        "operationId": "get_context_api_management_v1_context_get",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagementContextView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Context",
        "tags": [
          "management"
        ]
      }
    },
    "/api/management/v1/owner/default-companion": {
      "put": {
        "description": "Make one of my Eidolons the one that answers by default.\n\nPUT, so a phone that lost the response can simply ask again. Nothing\nelse moves: conversations already running keep the Eidolon they were\nstarted with, devices stay where they are, and no memory is copied —\nthis changes where *new* unaddressed work goes.",
        "operationId": "put_default_companion_api_management_v1_owner_default_companion_put",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DefaultCompanionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultCompanionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Default Companion",
        "tags": [
          "management"
        ]
      }
    }
  }
}