# server/eidolon_admin_server/local_api/management/router.py
"""The public Owner management surface: ``/api/management/v1``.

Its own module, not another block in ``app.py``. That file already mounts most
of the product and the plan's rule for this router is narrow enough to be worth
enforcing by shape: **authenticate, decode, call the backend, map the answer.**
No business judgement lives here, because this is the process that listens on
the LAN and deliberately holds no authority credential (plan §3.4.1) — a
decision made here would be a decision made on the wrong side of that boundary.

The Owner is never an input. It comes from the Controller session this router
authenticates, and is passed down as an argument.
"""

from __future__ import annotations

from typing import Literal, Protocol

from fastapi import APIRouter, Header, HTTPException, status
from pydantic import BaseModel, ConfigDict, Field

MANAGEMENT_PREFIX = "/api/management/v1"


class ManagementBackendError(RuntimeError):
    """The backend refused or could not answer; carries the status to relay."""

    def __init__(self, message: str, *, status_code: int = 503) -> None:
        super().__init__(message)
        self.status_code = status_code


class ManagementContextView(BaseModel):
    """What a client reads before it draws anything.

    ``capabilities`` is discovery, not permission: true means this Host can do
    the thing at all, and whether this Controller may is answered per action. A
    name absent from the map is one this client has never heard of — a version
    skew — while a name present and false is a feature this Host cannot do yet.

    ``limits`` values may be null, and a client must not substitute a number of
    its own: a limit nobody has measured is not a limit.
    """

    model_config = ConfigDict(extra="forbid")

    contract_version: Literal["1"] = "1"
    #: Deliberately no ``owner_id`` here — see the field note below.
    owner: "OwnerContextView"
    #: The Owner's pointer, named once. Null is a real state and no client may
    #: resolve it by choosing a Companion.
    default_companion_id: str | None = Field(default=None, max_length=64)
    capabilities: dict[str, bool]
    limits: dict[str, int | None]


class OwnerContextView(BaseModel):
    model_config = ConfigDict(extra="forbid")

    #: Returned because the client may need it for display and correlation, not
    #: because it may send one: a request that carried an ``owner_id`` would be
    #: asking this boundary to act for someone other than whoever it just
    #: authenticated, and no route here accepts one.
    owner_id: str = Field(min_length=1, max_length=64)
    display_name: str = Field(default="", max_length=128)
    #: The version a writer compares against when changing the default.
    revision: int = Field(ge=1)


class CompanionSummaryView(BaseModel):
    """One Eidolon, as a list shows it.

    No "is the default" flag. The page says which one is default exactly once,
    so a client cannot render two rows both claiming it, and cannot disagree
    with the Owner record about which one it is.
    """

    model_config = ConfigDict(extra="forbid")

    companion_id: str = Field(min_length=1, max_length=64)
    #: What the Owner named it. May be empty on a Host whose Data predates the
    #: name; a client shows its own placeholder rather than the identifier.
    display_name: str = Field(default="", max_length=128)
    #: Product type. A client must treat a value it does not know as "some
    #: other kind" and still render the row.
    kind: str = Field(min_length=1, max_length=32)
    #: Where it is in its life: active, retiring, archived, deleting. Four
    #: states rather than a boolean, because "the Owner archived it" and "it
    #: cannot run right now" are different things to show.
    lifecycle_state: str = Field(min_length=1, max_length=32)
    #: The version a later write compares against.
    revision: int = Field(ge=1)
    created_at: str = Field(min_length=1, max_length=64)
    updated_at: str = Field(min_length=1, max_length=64)


class CompanionRosterView(BaseModel):
    """One page of this Owner's Eidolons, oldest first."""

    model_config = ConfigDict(extra="forbid")

    contract_version: Literal["1"] = "1"
    #: Named once, for the page. Null is a real answer — every Companion
    #: archived, or the only one is a guard — and a client must show that state
    #: rather than promoting a row to fill the gap.
    default_companion_id: str | None = Field(default=None, max_length=64)
    companions: list[CompanionSummaryView]
    #: Opaque. A client stores it and sends it back to get the next page;
    #: parsing it would make the Host's page boundary part of the client.
    next_cursor: str | None = Field(default=None, max_length=256)


class CompanionDetailView(BaseModel):
    """One Eidolon, opened.

    ``is_default`` is here and *not* on a roster row on purpose. A single answer
    about a single Companion is a comparison the Host just made against the
    Owner's one pointer; a flag repeated across a list would let two rows claim
    it. Same fact, and only one shape of it can be self-contradictory.
    """

    model_config = ConfigDict(extra="forbid")

    contract_version: Literal["1"] = "1"
    companion_id: str = Field(min_length=1, max_length=64)
    display_name: str = Field(default="", max_length=128)
    kind: str = Field(min_length=1, max_length=32)
    lifecycle_state: str = Field(min_length=1, max_length=32)
    #: The version a later write must present. Read now so a client that is
    #: about to rename or archive is not made to fetch again first.
    revision: int = Field(ge=1)
    is_default: bool


class DefaultCompanionRequest(BaseModel):
    """Which of my Eidolons answers when I did not say which.

    ``expected_revision`` is not ceremony a client can skip: it is the version
    of the Owner it last read, and it is what makes two phones (or a phone and a
    browser) unable to both win. A client that has not read the Owner has no
    business changing this.
    """

    model_config = ConfigDict(extra="forbid")

    companion_id: str = Field(min_length=1, max_length=64)
    expected_revision: int = Field(ge=1)


class DefaultCompanionView(BaseModel):
    """Where the pointer now points, read back from the authority."""

    model_config = ConfigDict(extra="forbid")

    contract_version: Literal["1"] = "1"
    #: Echoed rather than assumed. A client that painted its own choice would be
    #: showing what it asked for instead of what happened.
    default_companion_id: str | None = Field(default=None, max_length=64)


class CompanionCreateRequest(BaseModel):
    """Ask for another Eidolon.

    ``operation_id`` is the client's, and it is what makes asking twice safe:
    every identifier the Host derives comes from it, so a retry addresses the
    same Eidolon rather than creating a second one. A client that generates a
    fresh id per attempt has opted out of that protection, which is why it is
    required rather than defaulted here.
    """

    model_config = ConfigDict(extra="forbid")

    operation_id: str = Field(min_length=36, max_length=36)
    display_name: str = Field(min_length=1, max_length=128)
    #: Absent means an ordinary conversational Eidolon. A client should not have
    #: to know the other values exist to create the normal thing.
    kind: str = Field(default="conversational", min_length=1, max_length=32)


class CompanionCreatedView(BaseModel):
    """The Eidolon that now exists, and whether its memory is up yet."""

    model_config = ConfigDict(extra="forbid")

    contract_version: Literal["1"] = "1"
    companion_id: str = Field(min_length=1, max_length=64)
    display_name: str = Field(default="", max_length=128)
    kind: str = Field(min_length=1, max_length=32)
    lifecycle_state: str = Field(min_length=1, max_length=32)
    revision: int = Field(ge=1)
    #: False when this request found the Eidolon already created by the same
    #: operation. It exists either way; a client that shows "created!" twice for
    #: one intent is telling the person something that did not happen.
    created: bool
    #: False means "its memory is not running yet", not "something is wrong".
    #: The Host converges on its own; a client should say "still coming up"
    #: rather than either hiding it or calling it a failure.
    memory_ready: bool


class ManagementBackendPort(Protocol):
    """What this router needs from the process that holds the credentials."""

    async def context(self, *, owner_id: str) -> dict: ...

    async def roster(self, *, owner_id: str, cursor: str | None) -> dict: ...

    async def companion(self, *, owner_id: str, companion_id: str) -> dict: ...

    async def set_default_companion(
        self, *, owner_id: str, companion_id: str, expected_revision: int
    ) -> dict: ...

    async def create_companion(
        self,
        *,
        owner_id: str,
        operation_id: str,
        display_name: str,
        kind: str,
    ) -> dict: ...


def register_management_routes(
    app,
    *,
    backend: ManagementBackendPort,
    authenticated_owner,
) -> None:
    """Mount the public management surface.

    ``authenticated_owner`` is supplied by the composition root: it verifies the
    Controller session and returns the Owner bound to it. Injected rather than
    imported so this module cannot reach for a different way to decide scope.
    """

    router = APIRouter(prefix=MANAGEMENT_PREFIX, tags=["management"])

    @router.get("/context", response_model=ManagementContextView)
    async def get_context(
        authorization: str | None = Header(default=None, alias="Authorization"),
    ) -> ManagementContextView:
        owner_id = await authenticated_owner(authorization)
        try:
            answer = await backend.context(owner_id=owner_id)
        except ManagementBackendError as exc:
            raise HTTPException(exc.status_code, str(exc)) from exc
        return ManagementContextView(
            owner=OwnerContextView(
                owner_id=answer["owner_id"],
                display_name=answer["owner_display_name"],
                revision=answer["owner_revision"],
            ),
            default_companion_id=answer["default_companion_id"],
            capabilities=answer["capabilities"],
            limits=answer["limits"],
        )

    @router.get("/companions", response_model=CompanionRosterView)
    async def list_companions(
        cursor: str | None = None,
        authorization: str | None = Header(default=None, alias="Authorization"),
    ) -> CompanionRosterView:
        """This Owner's Eidolons.

        There is no ``owner_id`` parameter and there will not be one: the Owner
        is whoever this session authenticated. ``cursor`` is the one thing a
        client may vary, and it is a value this boundary handed it.
        """
        owner_id = await authenticated_owner(authorization)
        try:
            answer = await backend.roster(owner_id=owner_id, cursor=cursor)
        except ManagementBackendError as exc:
            raise HTTPException(exc.status_code, str(exc)) from exc
        return CompanionRosterView(
            default_companion_id=answer["default_companion_id"],
            companions=[
                CompanionSummaryView(**row) for row in answer["companions"]
            ],
            next_cursor=answer["next_cursor"],
        )

    @router.get("/companions/{companion_id}", response_model=CompanionDetailView)
    async def get_companion(
        companion_id: str,
        authorization: str | None = Header(default=None, alias="Authorization"),
    ) -> CompanionDetailView:
        """One of this Owner's Eidolons.

        The Companion is in the path and the Owner is not expressible, so the
        pair is "this id, for whoever is signed in". One belonging to someone
        else answers 404, not 403: an id that answers differently can be probed
        for existence.
        """
        owner_id = await authenticated_owner(authorization)
        try:
            answer = await backend.companion(
                owner_id=owner_id, companion_id=companion_id
            )
        except ManagementBackendError as exc:
            raise HTTPException(exc.status_code, str(exc)) from exc
        return CompanionDetailView(
            companion_id=answer["companion_id"],
            display_name=answer["display_name"],
            kind=answer["kind"],
            lifecycle_state=answer["lifecycle_state"],
            revision=answer["revision"],
            is_default=answer["is_default"],
        )

    @router.put("/owner/default-companion", response_model=DefaultCompanionView)
    async def put_default_companion(
        payload: DefaultCompanionRequest,
        authorization: str | None = Header(default=None, alias="Authorization"),
    ) -> DefaultCompanionView:
        """Make one of my Eidolons the one that answers by default.

        PUT, so a phone that lost the response can simply ask again. Nothing
        else moves: conversations already running keep the Eidolon they were
        started with, devices stay where they are, and no memory is copied —
        this changes where *new* unaddressed work goes.
        """
        owner_id = await authenticated_owner(authorization)
        try:
            answer = await backend.set_default_companion(
                owner_id=owner_id,
                companion_id=payload.companion_id,
                expected_revision=payload.expected_revision,
            )
        except ManagementBackendError as exc:
            raise HTTPException(exc.status_code, str(exc)) from exc
        return DefaultCompanionView(
            default_companion_id=answer["default_companion_id"]
        )

    @router.put("/companions", response_model=CompanionCreatedView)
    async def create_companion(
        payload: CompanionCreateRequest,
        authorization: str | None = Header(default=None, alias="Authorization"),
    ) -> CompanionCreatedView:
        """Add another Eidolon for whoever is signed in.

        PUT with the client's operation id in the body rather than POST: the
        request states an end ("this operation has produced an Eidolon"), so a
        phone that lost the answer asks again and gets the same one. A POST here
        would make a second Eidolon the normal consequence of a dropped
        response.
        """
        owner_id = await authenticated_owner(authorization)
        try:
            answer = await backend.create_companion(
                owner_id=owner_id,
                operation_id=payload.operation_id,
                display_name=payload.display_name,
                kind=payload.kind,
            )
        except ManagementBackendError as exc:
            raise HTTPException(exc.status_code, str(exc)) from exc
        return CompanionCreatedView(
            companion_id=answer["companion_id"],
            display_name=answer["display_name"],
            kind=answer["kind"],
            lifecycle_state=answer["lifecycle_state"],
            revision=answer["revision"],
            created=answer["created"],
            memory_ready=answer["memory_ready"],
        )

    app.include_router(router)


ManagementContextView.model_rebuild()